Remove api file, incorporated into code
